chat_with_multiple_pdfs/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
langchain-community
altair
tiktoken
huggingface-hub

InstructorEmbedding==1.0.1
sentence-transformers==2.2.2 